fixed compiler errors

+++ exercises/1.3/software/project1/pathfinding/graph.h
@@ -84,7 +84,7 @@ void destroy_path_points(path_points* path);
 
 /* other */
 vertex* find_vertex_by_name(graph* graph, char* name);
-vertex* find_vertex_by_coords(graph* graph, float latitude, float longitude);
+vertex* find_vertex_by_coords(graph* graph, int latitude, int longitude);
 void draw_graph(graph* graph, int v_colour, int edge_colour);
 void draw_node(int colour, vertex* v);
 char** search_names(graph* graph, char* search_string, int* size);
